Move plugin JAR dependency to provided
The goal of specifying it at compile time was to pull Gerrit's versions
of dependencies when compiling. But this means the output JAR contains a
bunch of Gerrit classes, which we don't want. Instead, get transitive
dependencies from Gitiles and Soy, and exclude them from the shaded JAR.
This means we can't guarantee the versions we're compiling against are
binary compatible with the versions Gerrit is using at runtime, but
hopefully the version drift is manageable.

Change-Id: I98c2cc474611c83e2466978c71367627c2ed032a