Convert build to buck
Build a plugin jar linked that depends on this plugin being
cloned/linked into the Gerrit tree so it has access to Gerrit's buck
defs.

Build an output jar including resources in /+static. The
Gerrit-HttpStaticPrefix has to match the actual location in the jar of
the resources. Unlike the Gitiles dev server, which actually maps
/+static to an arbitrary resource path, we need to copy these to the
root of the jar.

Change-Id: I9974756cb6fdcf6efdf2effd62016431ea2d49c1