Add a link back to Gerrit on every page
It's convenient to have a link back to Gerrit in Gitiles.  This change
adds a "Code Review" link to the top-right menu of each page back to the
main Gerrit page.  This is similar to android.googlesource.com.

Gerrit's integrated Gitweb has "review" links back to the related
project home or change, but with the link in this change being in the
header, linking back to the Gerrit main page makes more sense.

Verified link shows while signed in and signed out on Gerrit v2.9.

Change-Id: Id1e3b67830eddec196c0d6215a9be4646141fa85