Update patch set 1
Patch Set 1: Code-Review+2

I was thinking about this recently and we could do something smarter, like trying to extract a change number/file path/etc. out of the URL and provide a more specific Gerrit link.

But I think this is a strict improvement, so let's get it submitted. Thanks!

Patch-set: 1
Label: Code-Review=+2