Update soy to newest version, and fix other problems
Upgrade guice multibindings to match version used by Gerrit. Package
JSR305 annotation, because closure-templates depends on it but Gerrit
doesn't package it anymore, as it was replaced by own class, not to
interfere with J2EE containers, in [1].

Note: Untill the new version of gitiles-servlet is published, a custom
built artifact must be used. The easiest way to achieve this is:

* increase VERSION in gitiles project, to say "0.1-3"
* run `buck build in` command in gitiles project
* Adjust this target in plugin's BUCK file:

maven_jar(
  name = 'gitiles-servlet',
  id = 'com.google.gitiles:gitiles-servlet:0.1-3',
  sha1 = '11dfaba6e8b7c392ae321b2c4b589001505f17ce',
  license = 'Apache2.0',
  repository = MAVEN_LOCAL,
  visibility = [],
)

* build the plugin as described in the documentation

[1] https://gerrit-review.googlesource.com/50112

Change-Id: I8f5fbdff0c45d3fc36e11314b27d83ae3f157a36