The plugin should supply links through WebLinks extensions
Today the Gitiles plug-in is using the [gitweb] support in
gerrit.config to add links to various places in the Gerrit UI.
This ensures that a restart is needed to enable gitiles plug-in
fully.

Gitiles should utilize the availible extension points to add
links to the UI and allow usage without extra configuration or
unneccessary init steps.

Change-Id: I3d2000fd616aebdd79fb0b5cfa7a9c06d0489254