Redirect to login if not authenticated
If a user isn't authenticated, redirect to login akin to the Gitweb
integration.  This avoids leaking information about repository
existence, but gives users a chance to log in if they have not yet.
This is especially useful as sessions expire: if the first page
viewed is Gitiles, they would receive a "Not Found" page instead of
creating a new session.

Change-Id: If06ec2a0ed85533b81a0e70d3545ee5129c7e3a6