Update patch set 3
Patch Set 3:

> The standalone build fails. It can't find the FileHistoryWeblink.
 > I built the api and installed it, the plugin api for 2.12-SNAPSHOT
 > seems to lie there good and ready but to no avail...
 > What am I missing?
 > 
 > [1]
 > .../gerrit-plugins/gitiles/src/main/java/com/googlesource/gerrit/plugins/gitiles/GitilesWeblinks.java:20:
 > error: cannot find symbol
 > import com.google.gerrit.extensions.webui.FileHistoryWebLink;

Buck is caching the API jar and you have to force it to get the updated one.

Try to set the sha1 of the jar to something else (any random value) in the BUCK file, then run the build again and it will fail with incorrect sha1. Then remove the sha1 (or replace with the correct one) and build again.

If it still doesn't work you'll have to take off and nuke ~/.gerritcodereview/buck-cache/cache/ from orbit.  That's the only way to be sure.

Patch-set: 3