Update patch set 4
Patch Set 4:

It should be possible to run f.i. GitBlit and gitiles simultaneously. It should be up to the admins to decide if they want to keep the old code browsing tool. It could be valid to have an option like:
Use gitiles as your default browser [Y/n]
But that wouldn't make sense if there are other code browsers using weblinks because they would still be there even if we remove [gitweb] section from Gerrit.config.
I don't think we can assume enough about the intentions of the installer to do something like that. If we could ascertain that the [gitweb] section is actually for gitiles than that could be valid, since you don't want duplicates.

Patch-set: 4