Prevent classloader issue when rendering markdown
Gitiles uses Grappa to render markdown. Grappa tries to load the GitilesMarkdown
parser from it's own or from the context classloader. This fails as Gitiles is
on the plugin's classloader only.

This change set's the plugin's classloader as the context classloader while the
filter chain is forwarded to the Gitiles servlet.

Change-Id: I060f0d53b3511f2bc3bd855e7e4c141acf749970