Continue ref search if first match is filtered out
A FilteredRepository is meant to act like a full repository with any
non-visible refs completely missing, so that a user cannot tell
between a ref that is inaccessible to them and a ref that doesn't
exist at all.

Unfortunately getRef, which is meant to resolve an abbreviated refname
along the standard search path, violates that.  Suppose I try to
getRef("foo") with the following refs present:

	refs/foo, present but inaccessible to me
	refs/heads/foo, present and accessible

Then the lookup finds refs/foo, which is correctly filtered out, and
the result is null.  But if refs/foo were not present, the result
would have been refs/heads/foo.  So by performing this lookup I would
discover that refs/foo exists.

Fix it by using exactRef to look up all possible matches along the
search path, filter by visibility, and return the first result.

Change-Id: I83f2bebcf1a4243f832f94661f9780fbc254806a