Use both global and per-repo config
As discussed in Change 71522, the global gitiles.conf should be
returned when per-repo config is unavailable. However, this was not
implemented in the past. This change deals with this issue and
automatically fallbacks to global gitiles.conf if the per-repo and
All-Projects gitiles.conf do not exist or are empty.

Change-Id: I5f99b92cfab0ea5d0f4b6c258d8ffa0e68cf2dee