Build with 2.14.9-SNAPSHOT
In upstream gerrit change I7af2f9f6c the commons-lang3 library
was removed from the plugin API. An explicit dependency was added
in this plugin by change I7120977be but this was only tested by
building the plugin in-tree.

Build with the 2.14.9-SNAPSHOT API in standalone mode to make sure
that everything works as expected when building in standalone mode.

Change-Id: Idb90cbb1cc6261429457e5cbccc9a09ebb758e4e