Install the MenuFilter as high priority on top of all URLs
There is no need to install the MenuFilter using the special mechanism of
gitiles.

The gitiles mechanism places the filter as low priority after the built-in
filters, so variables are not available for gitiles filters for rendering.

Installing the filter on top of all is simplicity and guarantee that the
banner settings are available in all cases.

Change-Id: I1c1da5186f8215395f7ee991512052f337967559
Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>