Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> What version of Gerrit has this been tested on?  It should probably go to one of the stable branches.

It was tested in master, it is a cleanup not a fix, but is required if we want to customize the not-found page if [1] is not accepted and we go at [2] route.

In any case it simplifies the handling, so I recommend of merging this anyway.

[1] https://gerrit-review.googlesource.com/c/gitiles/+/211312
[2] https://gerrit-review.googlesource.com/c/gitiles/+/213695

Patch-set: 2
Reviewer: Gerrit User 1040243 <1040243@173816e5-2b9a-37c3-8a2e-48639d4f1153>