Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> do we want to downport this (and the update to JGit 5.3.1) to stable-3.0 ?

There isn't a stable-3.0 branch on this plugin, and it has not advanced since stable-3.0 was cut on core.  So if we move the core jgit upgrade to stable-3.0 we can keep this change on master and explicitly set the revision.

The disadvantage is that if we merge this on master, core master will automatically advance the revision which will break the master build.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>