Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > I think we need a stable-3.0 branch on this plugin so that we don't get the recently merged changes from master on gerrit stable-3.0
> 
> But don't we follow the rule: "only create stable branch when really needed" also for core plugins? At least this applies also to codemirror-editor plugin.

In this case it is needed, if we want to get the merged in changes into the version of the plugin used on gerrit stable-3.0.  Otherwise we have to update to the plugin latest master revision, which will also include the other recently merged changes that we don't want there.

Patch-set: 1
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>