Add support for change clone repo link scheme
We use the HTTPS scheme for cloning the repository and at the same time SSH for bundling with Jenkins. Users want to see a link to the HTTPS scheme in the gitelis plugin interface.
If multiple schemes are specified in 'download.scheme' option, SSH scheme is used in clone repo link by default.

Our changes allow the administrator chooses the preferred clone repo link scheme.
The scheme is selected based on the 'download.scheme' option in the configuration file:
- If 'download.scheme' is not specified: arbitrarily prefer SSH, then HTTP, then git
- If multiple schemes are specified: prefer first listed

Change-Id: I57dafca482f771dd66241c1a27a55de3258d659f