Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> Is this similar to: [1]?
> 
> * [1] https://gerrit-review.googlesource.com/c/plugins/gitiles/+/235442

Yes. The author of the other CL proposed to abandon that one and move forward with this one. imo it makes more sense to set a single schema preference in configuration rather than setting an array. As long as that preference is valid (e.g. ssh is enabled), only that specific schema itself is succinct. If the first preference is not available, it's treated as invalid config and an error is logged in this CL.

Patch-set: 11
Reviewer: Gerrit User 1057569 <1057569@173816e5-2b9a-37c3-8a2e-48639d4f1153>