Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-2
> > 
> > Between 0.2-7 and 0.2-11 there are several library upgrades (jgit, soy, etc) so upgrading this in the gitiles plugin will make the library versions inconsistent with the versions used in core Gerrit.
> > 
> > If we want to upgrade this, we must also upgrade the corresponding libraries on core gerrit.
> 
> What about, instead, just including gitiles source code, similarly to what we did for Gerrit and JGit?
> We would then build gitiles with exactly the same libraries we have in Gerrit.
> 
> What do you think?

That will only work if the code on stable-0.2 is still compatible with the older versions of jgit, soy, etc that are use in core.

Maybe it would be easier to just upgrade those dependencies in core for the next 2.16 release?

If we want to include gitiles from source we could do it from stable-3.0 where it will be easier because the gitiles plugin is a core plugin.

Patch-set: 2
Reviewer: Gerrit User 1011123 <1011123@173816e5-2b9a-37c3-8a2e-48639d4f1153>