Upgrade gitiles-servlet to 0.2-7.1
Version 0.2-7 is quite old and does not include fixes that have
since been included in versions up to the latest 0.2-11 and on the
stable-0.2 branch.

In particular, [1] which was causing Gerrit JVM GC to start a series
of 'stop-the-world' when rendering large files, has been fixed in the
latest stable-0.2.

We cannot upgrade directly to 0.2-11 because that also includes updates
to JGit and soy which are not compatible with core Gerrit's 2.16 series.
Hence, 0.2-7.1 has been released which includes all of the bug fixes
and improvements, but not the library upgrades

[1] https://github.com/google/gitiles/issues/192

Change-Id: Iaad970f0400d14bd6c51fc1720aa12cc1b56c82a